+++ cpuprofile_test.go
@@ -24,7 +24,7 @@ func TestCPUProfile(t *testing.T) {
 	title := "cpuprofiletest"
 	cpuProfiler.StartProfiling(title)
 
-	_, err := ctx.CompileAndRun(profileScript, "script.js")
+	_, err := ctx.RunScript(profileScript, "script.js")
 	fatalIf(t, err)
 	val, err := ctx.Global().Get("start")
 	fatalIf(t, err)